# imo_vmdb/data/openapi.yaml
openapi: 3.1.0
info:
  title: imo-vmdb REST API
  version: "1"
  description: >
    Query rate and magnitude observations from the
    [Visual Meteor Database (VMDB)](https://www.imo.net/members/imo_vmdb/)
    stored in an *imo-vmdb* database.

    Full documentation: https://imo-vmdb.readthedocs.io/
servers:
  - url: /api/v1
    description: imo-vmdb web server (default port 8000)
paths:
  /rates:
    get:
      summary: Rate observations
      description: >
        Returns normalised rate observations.
        All filter parameters are optional and combined with AND.
        Multiple values for `shower`, `session_id`, and `rate_id` are
        combined with OR.

        Always returns the unpaginated total count in the `X-Total-Count`
        response header.  Use `limit=0` to obtain just the count.
      parameters:
        - $ref: '#/components/parameters/shower'
        - $ref: '#/components/parameters/period_start'
        - $ref: '#/components/parameters/period_end'
        - $ref: '#/components/parameters/sl_min'
        - $ref: '#/components/parameters/sl_max'
        - $ref: '#/components/parameters/lim_magn_min'
        - $ref: '#/components/parameters/lim_magn_max'
        - name: sun_alt_max
          in: query
          description: Maximum solar altitude in degrees.
          schema:
            type: number
        - name: moon_alt_max
          in: query
          description: Maximum lunar altitude in degrees.
          schema:
            type: number
        - $ref: '#/components/parameters/session_id'
        - name: rate_id
          in: query
          description: Filter by rate observation IDs (repeatable).
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - $ref: '#/components/parameters/rates_include'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: order_by
          in: query
          description: Sort column.
          schema:
            type: string
            enum: [id, period_start, period_end, sl_start, lim_magn]
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/fields'
      responses:
        "200":
          description: Successful response.
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatesResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /rates/{id}:
    get:
      summary: Single rate observation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateObservation'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /magnitudes:
    get:
      summary: Magnitude observations
      description: >
        Returns normalised magnitude observations.
        All filter parameters are optional and combined with AND.
        Multiple values for `shower`, `session_id`, and `magn_id` are
        combined with OR.

        Always returns the unpaginated total count in the `X-Total-Count`
        response header.  Use `limit=0` to obtain just the count.
      parameters:
        - $ref: '#/components/parameters/shower'
        - $ref: '#/components/parameters/period_start'
        - $ref: '#/components/parameters/period_end'
        - $ref: '#/components/parameters/sl_min'
        - $ref: '#/components/parameters/sl_max'
        - $ref: '#/components/parameters/lim_magn_min'
        - $ref: '#/components/parameters/lim_magn_max'
        - $ref: '#/components/parameters/session_id'
        - name: magn_id
          in: query
          description: Filter by magnitude observation IDs (repeatable).
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - $ref: '#/components/parameters/magnitudes_include'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: order_by
          in: query
          description: Sort column.
          schema:
            type: string
            enum: [id, period_start, period_end, sl_start, lim_magn]
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/fields'
      responses:
        "200":
          description: Successful response.
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagnitudesResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /magnitudes/{id}:
    get:
      summary: Single magnitude observation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagnitudeObservation'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /sessions:
    get:
      summary: Observation sessions
      description: >
        Returns observation sessions, optionally restricted to particular
        observers or to sessions that have at least one rate or magnitude
        observation matching the given filters.


        All observation-level filters (`shower`, `period_*`, `sl_*`,
        `lim_magn_*`) are evaluated via an EXISTS-join on `rate` and
        `magnitude`: a session is returned when at least one of its
        observations satisfies the criteria.  When `include=rates` or
        `include=magnitudes` is requested, the included observations
        are restricted by the same filters and to the session IDs in
        the response page.
      parameters:
        - name: observer_id
          in: query
          description: Filter by observer IDs (repeatable).
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - $ref: '#/components/parameters/shower'
        - $ref: '#/components/parameters/period_start'
        - $ref: '#/components/parameters/period_end'
        - $ref: '#/components/parameters/sl_min'
        - $ref: '#/components/parameters/sl_max'
        - $ref: '#/components/parameters/lim_magn_min'
        - $ref: '#/components/parameters/lim_magn_max'
        - $ref: '#/components/parameters/sessions_include'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: order_by
          in: query
          description: Sort column.
          schema:
            type: string
            enum: [id, country, observer_id]
        - $ref: '#/components/parameters/order'
      responses:
        "200":
          description: Successful response.
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionsResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /sessions/{id}:
    get:
      summary: Single observation session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /showers:
    get:
      summary: Shower reference data
      description: Returns all meteor showers from the reference data.
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shower'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /showers/active:
    get:
      summary: Showers active on a given date
      description: >
        Returns all showers whose activity period covers the given calendar
        date (year is ignored, year-wrapping showers handled correctly).
      parameters:
        - name: date
          in: query
          description: ISO date `YYYY-MM-DD`.  Defaults to today (UTC).
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shower'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /showers/{iau_code}:
    get:
      summary: Single meteor shower
      parameters:
        - name: iau_code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shower'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /showers/{iau_code}/radiants:
    get:
      summary: Radiant positions for a shower
      description: >
        Returns all radiant entries for the given shower, ordered by
        `(month, day)`.  Empty array if the shower has no radiants.
      parameters:
        - name: iau_code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Radiant'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /stats/meta:
    get:
      summary: Database scope summary
      description: >
        Returns the total number of sessions, rates and magnitudes plus the
        covered date range.  Useful for clients to discover what's in the
        database before issuing more specific queries.
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsMeta'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /stats/by-shower:
    get:
      summary: Per-shower aggregate counts
      parameters:
        - $ref: '#/components/parameters/period_start'
        - $ref: '#/components/parameters/period_end'
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShowerStat'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /stats/by-country:
    get:
      summary: Per-country aggregate counts
      parameters:
        - $ref: '#/components/parameters/period_start'
        - $ref: '#/components/parameters/period_end'
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryStat'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
  /stats/by-year:
    get:
      summary: Per-year aggregate counts
      parameters:
        - $ref: '#/components/parameters/period_start'
        - $ref: '#/components/parameters/period_end'
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YearStat'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          $ref: '#/components/responses/NoDB'
...
components:
  parameters:
    shower:
      name: shower
      in: query
      description: >
        IAU shower code(s) (repeatable).
        Use `SPO` to include sporadic meteors (returned as `null` in the
        `shower` field).
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      examples:
        single:
          value: [PER]
        multiple:
          value: [PER, GEM]
        with_sporadic:
          value: [PER, SPO]
    period_start:
      name: period_start
      in: query
      description: >
        Lower bound for `period_start`.  Strict ISO 8601:
        `YYYY-MM-DDTHH:MM:SS` (e.g. `2015-08-12T20:00:00`).  IMO data
        is UTC by convention; no timezone marker is accepted.
        Space-separator, date-only, trailing `Z` and `±HH:MM` offsets
        are rejected with HTTP 400.
      schema:
        type: string
        format: date-time
        example: "2015-08-12T20:00:00"
    period_end:
      name: period_end
      in: query
      description: >
        Upper bound for `period_end`.  Strict ISO 8601:
        `YYYY-MM-DDTHH:MM:SS`.  Same rules as `period_start` (no
        timezone marker, no space separator, no date-only).
      schema:
        type: string
        format: date-time
        example: "2015-08-31T23:59:59"
    sl_min:
      name: sl_min
      in: query
      description: Minimum solar longitude (degrees, `sl_start >=`).
      schema:
        type: number
        minimum: 0
        maximum: 360
    sl_max:
      name: sl_max
      in: query
      description: Maximum solar longitude (degrees, `sl_end <=`).
      schema:
        type: number
        minimum: 0
        maximum: 360
    lim_magn_min:
      name: lim_magn_min
      in: query
      description: Minimum limiting magnitude.
      schema:
        type: number
    lim_magn_max:
      name: lim_magn_max
      in: query
      description: Maximum limiting magnitude.
      schema:
        type: number
    session_id:
      name: session_id
      in: query
      description: Filter by session IDs (repeatable).
      schema:
        type: array
        items:
          type: integer
      style: form
      explode: true
...
    limit:
      name: limit
      in: query
      description: >
        Maximum number of rows to return.  Use `limit=0` to obtain only the
        total count via the `X-Total-Count` header.  Without `limit`, all
        matching rows are returned (backwards-compatible behaviour).
      schema:
        type: integer
        minimum: 0
    offset:
      name: offset
      in: query
      description: Number of leading rows to skip.
      schema:
        type: integer
        minimum: 0
...